Added score tracking.

+++ Assets/Scripts/FaceMaker/FaceTextureStore.cs
@@ -11,7 +11,7 @@ public class FaceTextureStore : MonoBehaviour
     public Texture2D BlueVassalFace;
     public Texture2D RedVassalFace;
 
-    public int FinalScore;
+    public long FinalScore;
 
     public bool Old = false;
